TINY-11394: refactor
